reuse _parse_labeled_data function
