+++ learntorank/_modidx.py
@@ -38,8 +38,6 @@ d = { 'settings': { 'branch': 'main',
                                                                                                   'learntorank/evaluation.py'),
                                         'learntorank.evaluation._evaluate_query_retry': ( 'module_evaluation.html#_evaluate_query_retry',
                                                                                           'learntorank/evaluation.py'),
-                                        'learntorank.evaluation._parse_labeled_data': ( 'module_evaluation.html#_parse_labeled_data',
-                                                                                        'learntorank/evaluation.py'),
                                         'learntorank.evaluation.evaluate': ('module_evaluation.html#evaluate', 'learntorank/evaluation.py'),
                                         'learntorank.evaluation.evaluate_query': ( 'module_evaluation.html#evaluate_query',
                                                                                    'learntorank/evaluation.py')},
@@ -128,6 +126,8 @@ d = { 'settings': { 'branch': 'main',
                                                                                        'learntorank/query.py'),
                                    'learntorank.query._annotate_data': ('module_query.html#_annotate_data', 'learntorank/query.py'),
                                    'learntorank.query._build_query_body': ('module_query.html#_build_query_body', 'learntorank/query.py'),
+                                   'learntorank.query._parse_labeled_data': ( 'module_query.html#_parse_labeled_data',
+                                                                              'learntorank/query.py'),
                                    'learntorank.query.collect_vespa_features': ( 'module_query.html#collect_vespa_features',
                                                                                  'learntorank/query.py'),
                                    'learntorank.query.send_query': ('module_query.html#send_query', 'learntorank/query.py'),
